requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
croniter==6.0.0
Django==5.2.1
django-picklefield==3.3
django-q2==1.8.0
flake8==7.2.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
